cleanup

+++ ptychodus/api/apparatus.py
@@ -1,19 +1,18 @@
 from dataclasses import dataclass
-from decimal import Decimal
 
 
 @dataclass(frozen=True)
 class FresnelZonePlate:
-    zonePlateDiameterInMeters: Decimal
-    outermostZoneWidthInMeters: Decimal
-    centralBeamstopDiameterInMeters: Decimal
+    zonePlateDiameterInMeters: float
+    outermostZoneWidthInMeters: float
+    centralBeamstopDiameterInMeters: float
 
-    def focalLengthInMeters(self, centralWavelengthInMeters: Decimal) -> Decimal:
+    def focalLengthInMeters(self, centralWavelengthInMeters: float) -> float:
         return self.zonePlateDiameterInMeters * self.outermostZoneWidthInMeters \
                 / centralWavelengthInMeters
 
 
 @dataclass(frozen=True)
 class PixelGeometry:
-    widthInMeters: Decimal
-    heightInMeters: Decimal
+    widthInMeters: float
+    heightInMeters: float
